requirements.txt
amqp==5.1.1
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.1
django-debug-toolbar==3.7.0
django-debug-toolbar-force==0.1.8
django-extensions==3.2.1
django-filter==22.1
django-jazzmin==2.6.0
django-nine==0.2.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
executing==1.0.0
factory-boy==3.2.1
Faker==14.2.0
flake8==5.0.4
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
ipython==8.5.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
oauthlib==3.2.1
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pluggy==1.0.0
prompt-toolkit==3.0.31
psycopg2==2.9.5
psycopg2-binary==2.9.3
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.5.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
stack-data==0.5.0
tomli==2.0.1
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.2
Unidecode==1.3.4
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
